Move Logger class
