+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/util/IOUtils.java
@@ -8,6 +8,8 @@ import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.toolslab.quickcode.util.log.Logger;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
