+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/util/UriUtils.java
@@ -12,6 +12,8 @@ import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.toolslab.quickcode.util.log.Logger;
+
 import java.io.BufferedReader;
 import java.io.FileDescriptor;
 import java.io.FileNotFoundException;
